batch rendering live vehicles to map
