+++ src/stores/live_track_store.ts
@@ -2,6 +2,7 @@ import { get, writable } from 'svelte/store';
 import type { Writable } from 'svelte/store';
 import type { CircleMarker, Map } from 'leaflet';
 import L from 'leaflet';
+import { handleAddOrUpdateEvent } from '../explorer-api/batch_live_updates';
 
 // Define the shape of vehicle attributes
 interface VehicleAttributes {
@@ -13,13 +14,13 @@ interface VehicleAttributes {
 }
 
 // Define the shape of a vehicle object
-interface Vehicle {
+export interface Vehicle {
   id: string;
   attributes: VehicleAttributes;
 }
 
 // Define the state for each vehicle
-interface VehicleStateMap {
+export interface VehicleStateMap {
   marker: CircleMarker;
   data: VehicleAttributes;
 }
@@ -27,13 +28,13 @@ interface VehicleStateMap {
 // Store for tracking vehicle state
 export const vehicleStateMap: Writable<Record<string, VehicleStateMap>> = writable({});
 
-export function isDataFresh(vehicle: any): boolean {
-  const { id, attributes } = vehicle;
-  const currentState = get(vehicleStateMap);
-  const existingData = currentState[id]?.data;
+// export function isDataFresh(vehicle: any): boolean {
+//   const { id, attributes } = vehicle;
+//   const currentState = get(vehicleStateMap);
+//   const existingData = currentState[id]?.data;
 
-  return !existingData || new Date(attributes.updated_at) > new Date(existingData.updated_at);
-}
+//   return !existingData || new Date(attributes.updated_at) > new Date(existingData.updated_at);
+// }
 
 // Handle the reset event
 export function handleResetEvent(data: Vehicle[], map: Map) {
@@ -46,42 +47,42 @@ export function handleResetEvent(data: Vehicle[], map: Map) {
   data.forEach((vehicle) => handleAddOrUpdateEvent(vehicle, map));
 }
 
-// Handle add or update event
-export function handleAddOrUpdateEvent(vehicle: Vehicle, map: Map) {
-  if (!isDataFresh(vehicle)) {
-    return
-  }
+// // Handle add or update event
+// export function handleAddOrUpdateEvent(vehicle: Vehicle, map: Map) {
+//   if (!isDataFresh(vehicle)) {
+//     return
+//   }
 
-  const { id, attributes } = vehicle;
+//   const { id, attributes } = vehicle;
 
-  if (!id) {
-    console.error("Vehicle ID is missing");
-    return;
-  }
+//   if (!id) {
+//     console.error("Vehicle ID is missing");
+//     return;
+//   }
 
-  vehicleStateMap.update((state) => {
-    const existingVehicle = state[id];
+//   vehicleStateMap.update((state) => {
+//     const existingVehicle = state[id];
 
-    if (existingVehicle) {
-      // Update existing marker
-      const { marker } = existingVehicle;
-      marker.setLatLng([attributes.latitude, attributes.longitude]);
-      state[id].data = attributes; // Update state data
-    } else {
-      // Add new marker
-      const marker = L.circle([attributes.latitude, attributes.longitude], {
-        color: "white",
-        fillColor: "white",
-        fillOpacity: 1,
-        radius: 60,
-      }).addTo(map);
+//     if (existingVehicle) {
+//       // Update existing marker
+//       const { marker } = existingVehicle;
+//       marker.setLatLng([attributes.latitude, attributes.longitude]);
+//       state[id].data = attributes; // Update state data
+//     } else {
+//       // Add new marker
+//       const marker = L.circle([attributes.latitude, attributes.longitude], {
+//         color: "white",
+//         fillColor: "white",
+//         fillOpacity: 1,
+//         radius: 60,
+//       }).addTo(map);
 
-      state[id] = { marker, data: attributes };
-    }
+//       state[id] = { marker, data: attributes };
+//     }
 
-    return state;
-  });
-}
+//     return state;
+//   });
+// }
 
 export function handleRemoveEvent(vehicle: Vehicle) {
   const { id } = vehicle;
